Related work items: #119, #120, #121, #71, #95

+++ Drugstore.Core/ExternalDrugstore/ExternalDrugstoreMedicine.cs
@@ -14,6 +14,6 @@ namespace Drugstore.Core
         [Required]
         public MedicineOnStock StockMedicine { get; set; }
         public uint Quantity { get; set; }
-        public float PricePerOne { get; set; }
+        public double PricePerOne { get; set; }
     }
 }
