+++ Drugstore/Models/Patient/PatientTreatmentViewModel.cs
@@ -10,7 +10,7 @@ namespace Drugstore.Models
         public int Id { get; set; }
         public int CurrentPage { get; set; }
         public int TotalPages { get; set; }
-        public float TotalCost { get; set; }
+        public double TotalCost { get; set; }
         public List<PrescriptionGeneralDataModel> Prescriptions { get; set; }
         public bool IsValid { get; set; } = true;
         public string Error { get; set; } = "";
